requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.2
yfinance==0.2.43
plotly==5.23.0
scipy==1.13.0
matplotlib==3.9.0
matplotlib-inline==0.1.6
pymongo==4.8.0
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langgraph==0.2.25
langchain-groq==0.1.9
duckduckgo-search==6.3.2
groq==0.11.0
